+++ crates/plotinator-plot-ui/src/plot_settings/loaded_logs.rs
@@ -1,6 +1,6 @@
-use chrono::NaiveDateTime;
-use egui::{Color32, Key, RichText, TextEdit};
+use egui::{Color32, Key, RichText};
 use egui_phosphor::regular;
+use egui_plot::PlotBounds;
 use plotinator_strfmt::format_data_size;
 
 use crate::WARN_ON_UNPARSED_BYTES_THRESHOLD;
@@ -61,25 +61,28 @@ fn ui_button_open_log(ui: &mut egui::Ui, loaded_log: &mut LoadedLogSettings) {
     }
 }
 
-pub fn show_log_date_settings_ui(ui: &mut egui::Ui, loaded_log: &mut LoadedLogSettings) {
+pub fn show_log_settings_ui(
+    ui: &mut egui::Ui,
+    loaded_log: &mut LoadedLogSettings,
+    selected_box: Option<PlotBounds>,
+) {
     ui_button_open_log(ui, loaded_log);
     ui.label("");
     ui.horizontal(|ui| {
         ui_button_toggle_visibility(ui, loaded_log);
         ui_button_remove_toggle(ui, loaded_log);
     });
-    if loaded_log.tmp_date_buf.is_empty() {
-        loaded_log.tmp_date_buf = loaded_log
-            .start_date()
-            .format("%Y-%m-%d %H:%M:%S%.f")
-            .to_string();
-    }
     if loaded_log.clicked() {
-        log_settings_window(ui, loaded_log, &loaded_log.log_label());
+        log_settings_window(ui, loaded_log, &loaded_log.log_label(), selected_box);
     }
 }
 
-fn log_settings_window(ui: &egui::Ui, settings: &mut LoadedLogSettings, log_name_date: &str) {
+fn log_settings_window(
+    ui: &egui::Ui,
+    settings: &mut LoadedLogSettings,
+    log_name_date: &str,
+    selected_box: Option<PlotBounds>,
+) {
     // State of window bound to the 'X'-button that closes the window
     let mut open = true;
     egui::Window::new(RichText::new(log_name_date).size(20.0).strong())
@@ -99,35 +102,18 @@ fn log_settings_window(ui: &egui::Ui, settings: &mut LoadedLogSettings, log_name
                         ui.label("Modify the start date to offset the plots of this log");
                         ui.label(format!("original date: {}", settings.original_start_date));
                         ui.label(RichText::new("YYYY-mm-dd HH:MM:SS.ms").strong());
-                        let date_txt_input_resp =
-                            ui.add(TextEdit::singleline(&mut settings.tmp_date_buf));
-                        date_txt_input_resp.request_focus();
-                        if date_txt_input_resp.changed() {
-                            match NaiveDateTime::parse_from_str(
-                                &settings.tmp_date_buf,
-                                "%Y-%m-%d %H:%M:%S%.f",
-                            ) {
-                                Ok(new_dt) => {
-                                    settings.err_msg.clear();
-                                    settings.new_date_candidate = Some(new_dt);
-                                }
-                                Err(e) => {
-                                    settings.err_msg = format!("⚠ {e} ⚠");
-                                }
-                            };
+                        let date_txt_input_resp = settings.start_date_editor.show(ui);
+                        if !settings.log_points_cutter.clicked {
+                            date_txt_input_resp.request_focus();
                         }
-                        if settings.err_msg.is_empty() {
-                            if let Some(new_date) = settings.new_date_candidate {
-                                if ui.button("Apply").clicked()
-                                    || ui.input(|i| i.key_pressed(Key::Enter))
-                                {
-                                    settings.new_start_date(new_date.and_utc());
-                                    settings.date_changed = true;
-                                    log::info!("New date: {}", settings.start_date());
-                                }
+
+                        if let Some(new_date) = settings.start_date_editor.current() {
+                            if ui.button("Apply").clicked()
+                                || ui.input(|i| i.key_pressed(Key::Enter))
+                            {
+                                settings.new_start_date(new_date);
+                                log::info!("New date: {}", settings.start_date());
                             }
-                        } else {
-                            ui.label(settings.err_msg.clone());
                         }
                         if ui.button("Cancel").clicked() {
                             *settings.clicked_mut() = false;
@@ -141,6 +127,9 @@ fn log_settings_window(ui: &egui::Ui, settings: &mut LoadedLogSettings, log_name
                 }
             });
 
+            if ui.button("Cut Points").clicked() {
+                settings.log_points_cutter.clicked = true;
+            }
             if let Some(log_metadata) = settings.log_metadata() {
                 ui.separator();
                 egui::ScrollArea::vertical()
@@ -155,8 +144,15 @@ fn log_settings_window(ui: &egui::Ui, settings: &mut LoadedLogSettings, log_name
             }
         });
 
+    if settings.log_points_cutter.clicked {
+        settings
+            .log_points_cutter
+            .show(ui, log_name_date, selected_box);
+    }
+
     if !open || ui.ctx().input(|i| i.key_pressed(Key::Escape)) {
         *settings.clicked_mut() = false;
+        settings.log_points_cutter.clicked = false;
     }
 }